requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
joblib
PyPDF2
pypdf
google-genai
python-dotenv
streamlit-lottie
requests
# Interview feature dependencies
fastapi
uvicorn[standard]
websockets